smart_content_engine/requirements.txt
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0
requests-oauthlib==1.3.1
schedule==1.2.0
beautifulsoup4==4.12.2
aiofiles==23.2.0